+++ integrations/intellij/gradle.properties
@@ -14,8 +14,8 @@ pluginUntilBuild = 232
 # See https://jb.gg/intellij-platform-builds-list for available build versions.
 pluginVerifierIdeVersions = 2022.2,2022.3,2023.1
 
-platformType = IC
-platformVersion = 2022.2
+platformType = IU
+platformVersion = 2023.1
 platformDownloadSources = true
 
 # Plugin Dependencies -> https://plugins.jetbrains.com/docs/intellij/plugin-dependencies.html
