--- integrations/intellij/src/main/kotlin/com/previewjs/intellij/plugin/listeners/ProjectManagerListener.kt
@@ -1,12 +1,0 @@
-package com.previewjs.intellij.plugin.listeners
-
-import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
-import com.intellij.openapi.project.ProjectManagerListener
-import com.previewjs.intellij.plugin.services.ProjectService
-
-internal class ProjectManagerListener : ProjectManagerListener {
-    override fun projectOpened(project: Project) {
-        project.service<ProjectService>()
-    }
-}
