+++ integrations/intellij/src/main/resources/META-INF/plugin.xml
@@ -16,27 +16,7 @@
     <depends>com.intellij.modules.platform</depends>
 
     <extensions defaultExtensionNs="com.intellij">
-        <projectService serviceImplementation="com.previewjs.intellij.plugin.services.ProjectService"/>
         <notificationGroup id="Preview.js" displayType="BALLOON" />
-
-<!--
-
-IntelliJ Community Edition language list: JShellLanguage, SHRINKER_CONFIG, HTML, RELAX-NG, JSON5, JAVA, yaml, TOML, DeviceSpec, textmate, TEXT, kotlin, LogcatFilter, JQL, RoomSql, GitExclude, AndroidDataBinding, HgIgnore, IntegrationPerformanceTest, Lombok.Config, EditorConfig, JSON Lines, Properties, IgnoreLang, UAST, XPath2, AIDL, SPI, MultiDexKeep, KND, HtmlCompatible, XPath, AGSL, GitIgnore, DTD, , JSON, XHTML, Groovy, Markdown, Renderscript, Manifest, JVM, RegExp, XML, SVG, Shell Script, Smali
-
-Generated with:
-
-Language.getRegisteredLanguages().joinToString(", ") { l -> l.id }
--->
-
-<!--        <codeInsight.inlayProvider language="IgnoreLang" implementationClass="com.previewjs.intellij.plugin.inlays.InlayProvider" />-->
-<!--        <codeInsight.inlayProvider language="TypeScript JSX" implementationClass="com.previewjs.intellij.plugin.inlays.InlayProvider" />-->
-
         <codeInsight.inlayProviderFactory implementation="com.previewjs.intellij.plugin.inlays.InlayProviderFactory" />
-
     </extensions>
-
-    <applicationListeners>
-        <listener class="com.previewjs.intellij.plugin.listeners.ProjectManagerListener"
-                  topic="com.intellij.openapi.project.ProjectManagerListener"/>
-    </applicationListeners>
 </idea-plugin>
